extract history module from page router
